+++ aixcalibuha/examples/sen_analysis_example.py
@@ -3,13 +3,13 @@ Example file for the senanalyzer package. The usage of modules and classes insid
 the senanalyzer package should be clear when looking at the examples.
 If not, please raise an issue.
 """
-
-from ebcpy.examples import dymola_api_example
+import os
+from ebcpy import DymolaAPI
 from aixcalibuha.sensanalyzer import SobolAnalyzer
 from aixcalibuha.examples import cal_classes_example
 
 
-def example_sensitivity_analysis(sim_api, cal_classes, stat_measure):
+def example_sensitivity_analysis(sim_api, cal_classes):
     """
     Example process of a sensitivity analysis.
     First, the sensitivity problem is constructed, in this example
@@ -25,9 +25,6 @@ def example_sensitivity_analysis(sim_api, cal_classes, stat_measure):
     :param list cal_classes:
         List of :meth:`calibration-class<aixcalibuha.data_types.CalibrationClass>`
         objects to be analyzed.
-    :param str stat_measure:
-        The statistical measure, one of the possible of
-        the :meth:`statistics analyzer<ebcpy.utils.statistics_analyzer.StatisticsAnalyzer>`
     :return: A list calibration classes
     :rtype: list
     """
@@ -35,7 +32,6 @@ def example_sensitivity_analysis(sim_api, cal_classes, stat_measure):
 
     sen_analyzer = SobolAnalyzer(
             sim_api=sim_api,
-            statistical_measure=stat_measure,
             num_samples=1,
             cd=sim_api.cd,
             analysis_variable='S1'
@@ -52,12 +48,10 @@ def example_sensitivity_analysis(sim_api, cal_classes, stat_measure):
 
 if __name__ == "__main__":
     # Parameters for sen-analysis:
-    STATISTICAL_MEASURE = "RMSE"
-
-    DYM_API = dymola_api_example.setup_dymola_api()
+    DYM_API = DymolaAPI(model_name="",
+                        cd=os.getcwd())
     CALIBRATION_CLASSES = cal_classes_example.setup_calibration_classes()
 
     # Sensitivity analysis:
     CALIBRATION_CLASSES = example_sensitivity_analysis(DYM_API,
-                                                       CALIBRATION_CLASSES,
-                                                       STATISTICAL_MEASURE)
+                                                       CALIBRATION_CLASSES)